Performance and diff2html API Rework (#18)
* Fix a bug with diff hunk header value optionality

* Optimizing exceptionally bad diffs.

* Improve DiffRenderer API

* Renamed to DiffHunksRenderer (since it'll make sense as we roll out
  api updates)
* Added configurability to a bunch of the options

* Make line numbers configurable

* First pass at new `DiffFileRenderer` class

This is a foundation to base API in a somewhat similar way to diff2HTML.

This encapsulates some of the functionality that i was manually putting
together into main.

It also starts to productize the file based API a bit. Obviously next
steps would be a Diff class for full diffs, that could or could not
optionally render metadata and stuff.

* Make the pathc fetcher async

No point in loading the massive patch files for this demo app unless its
absolutely needed

* useless variable name rewrites

* smol loading improvements

Ensure we don't double fetch, no reason tbh